+++ src/contexts/Validators/Utils.ts
@@ -1,8 +1,10 @@
 // Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+import type BigNumber from 'bignumber.js';
+import type { LocalMeta } from 'contexts/FastUnstake/types';
 import type { LocalExposureData, Validator } from 'contexts/Validators/types';
-import type { NetworkName } from 'types';
+import type { AnyJson, NetworkName } from 'types';
 
 // Get favorite validators from local storage.
 export const getLocalFavorites = (network: NetworkName) => {
@@ -38,4 +40,55 @@ export const setEraLocalExposures = (
       },
     })
   );
+};
+
+// Validate local exposure metadata.
+export const validateLocalExposure = (
+  localMeta: AnyJson,
+  endEra: BigNumber
+): LocalMeta | null => {
+  const localIsExposed = localMeta?.isExposed ?? null;
+  let localChecked = localMeta?.checked ?? null;
+
+  // check types saved.
+  if (typeof localIsExposed !== 'boolean' || !Array.isArray(localChecked)) {
+    return null;
+  }
+  // check checked only contains numbers.
+  const checkedNumeric = localChecked.every((e) => typeof e === 'number');
+  if (!checkedNumeric) {
+    return null;
+  }
+
+  // remove any expired eras and sort highest first.
+  localChecked = localChecked
+    .filter((e: number) => endEra.isLessThan(e))
+    .sort((a: number, b: number) => b - a);
+
+  // if no remaining eras, invalid.
+  if (!localChecked.length) {
+    return null;
+  }
+
+  // check if highest -> lowest are decremented, no missing eras.
+  let i = 0;
+  let prev = 0;
+  const noMissingEras = localChecked.every((e: number) => {
+    i++;
+    if (i === 1) {
+      prev = e;
+      return true;
+    }
+    const p = prev;
+    prev = e;
+    if (e === p - 1) return true;
+    return false;
+  });
+  if (!noMissingEras) {
+    return null;
+  }
+  return {
+    isExposed: localIsExposed,
+    checked: localChecked,
+  };
 };
